remove general.py
