+++ ultralytics/yolo/data/augment.py
@@ -8,7 +8,10 @@ import numpy as np
 import torch
 import torchvision.transforms as T
 
-from ..utils.general import LOGGER, check_version, colorstr, segment2box
+from ..utils import LOGGER
+from ..utils.checks import check_version
+from ..utils.loggers import colorstr
+from ..utils.ops import segment2box
 from ..utils.instance import Instances
 from ..utils.metrics import bbox_ioa
 from .utils import IMAGENET_MEAN, IMAGENET_STD, polygons2masks, polygons2masks_overlap
