+++ ultralytics/yolo/data/base.py
@@ -9,7 +9,7 @@ import numpy as np
 from torch.utils.data import Dataset
 from tqdm import tqdm
 
-from ..utils.general import NUM_THREADS
+from ..utils import NUM_THREADS
 from .utils import BAR_FORMAT, HELP_URL, IMG_FORMATS, LOCAL_RANK
