+++ ultralytics/yolo/data/dataset.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import torchvision
 from tqdm import tqdm
 
-from ..utils.general import NUM_THREADS
+from ..utils import NUM_THREADS
 from .augment import *
 from .base import BaseDataset
 from .utils import BAR_FORMAT, HELP_URL, LOCAL_RANK, get_hash, img2label_paths, verify_image_label
