+++ ultralytics/yolo/utils/checks.py
@@ -1,4 +1,16 @@
-from .general import LOGGER
+from pathlib import Path
+import platform
+import pkg_resources as pkg
+from subprocess import check_output
+import urllib
+import glob
+import sys
+import os
+
+import torch
+from ultralytics.yolo.utils import ROOT, LOGGER, CONFIG_DIR, FONT, AUTOINSTALL, TryExcept
+from .loggers import colorstr, emojis
+
 
 
 def check_version(current="0.0.0", minimum="0.0.0", name="version ", pinned=False, hard=False, verbose=False):
