+++ ultralytics/yolo/utils/instance.py
@@ -5,7 +5,7 @@ from typing import List
 
 import numpy as np
 
-from .general import ltwh2xywh, ltwh2xyxy, resample_segments, xywh2ltwh, xywh2xyxy, xyxy2ltwh, xyxy2xywh
+from .ops import ltwh2xywh, ltwh2xyxy, resample_segments, xywh2ltwh, xywh2xyxy, xyxy2ltwh, xyxy2xywh
 
 
 # From PyTorch internals
